% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{accuracy}
\alias{accuracy}
\title{Compute predictive accuracy.}
\usage{
accuracy(x, y)
}
\arguments{
\item{x}{A vector of predicted labels.}

\item{y}{A vector of true labels.}
}
\description{
Compute predictive accuracy.
}
\examples{
data(car)
dpl.lshade <- lshade(NP = 20, G = 25, data = car, class.name = names(car)[7], c = 0.1,
structure = "tan", pB = 0.05, edgelist = NULL, verbose = 5)
p <- predict(dpl.lshade$Best, car)
accuracy(p, car$class)
}
