% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_edgelist}
\alias{apply_edgelist}
\title{Apply a custom edgelist to a bnclassify BN structure}
\usage{
apply_edgelist(bn, edgelist, class.name)
}
\arguments{
\item{bn}{A bnc_dag object from bnclassify}

\item{edgelist}{A matrix with 2 columns (from, to) representing directed edges}

\item{class.name}{The name of the class variable}
}
\description{
Apply a custom edgelist to a bnclassify BN structure
}
\keyword{internal}
