% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastCLL}
\alias{fastCLL}
\title{Fast Conditional Log-Likelihood without DAG verification}
\usage{
fastCLL(net, cll_data)
}
\arguments{
\item{net}{A bnc_bn object}

\item{cll_data}{Precomputed data from precompute_cll_data}
}
\description{
Fast Conditional Log-Likelihood without DAG verification
}
\keyword{internal}
