% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_family}
\alias{get_family}
\title{Retrieve Families for a Given Node in Bayesian Network}
\usage{
get_family(node, df, class.str)
}
\arguments{
\item{node}{A node in the bayesian network.}

\item{df}{An adjacency list representing the network, formatted as a data frame.}

\item{class.str}{The name of the class node.}
}
\value{
A vector containing the node and its family members.
}
\description{
This function identifies and returns the family for a given node in a Bayesian network.
The family is a set of nodes, which includes the node itself, its parents, and the class.
}
\keyword{internal}
