% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPBest}
\alias{getPBest}
\title{JADE uses a mutation strategy called ”DE/current-to- p best”
Vectors are selected from the best p vectors in the current
population}
\usage{
getPBest(x, n = 30)
}
\description{
JADE uses a mutation strategy called ”DE/current-to- p best”
Vectors are selected from the best p vectors in the current
population
}
\keyword{internal}
