% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{individual}
\alias{individual}
\title{Generates an individual (potential solution)}
\usage{
individual(W, X, Y, Z)
}
\arguments{
\item{W}{Number of nodes in BN structure}

\item{X}{Cardinality of CPTs}

\item{Y}{Number of parameters per node}

\item{Z}{BN Structure with CPT slots}
}
\description{
Generates an individual (potential solution)
}
\keyword{internal}
