% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jso.R
\name{jso}
\alias{jso}
\title{Discriminative parameter learning of BN by jSO.}
\usage{
jso(
  NP = 40,
  G = 100,
  data,
  class.name,
  structure = c("nb", "tancl", "tan", "hc"),
  edgelist = NULL,
  verbose = 25,
  ...
)
}
\arguments{
\item{NP}{positive integer giving the number of candidate solutions in the initial population.}

\item{G}{positive integer specifying the maximum number of generations.}

\item{data}{The data frame from which to learn the classifier.}

\item{class.name}{A character. Name of the class variable.}

\item{structure}{A character. Name of the structure learning function. "tan" uses Tree Augmented Network.
"nb" uses Naive Bayes. "hc" uses Hill Climbing.}

\item{edgelist}{A matrix. An optional edge list to use a custom BN structure.}

\item{verbose}{positive integer indicating the number of generations until the iteration progress should be printed.}

\item{...}{other structure learning options from \link[bnclassify]{tan_cl} or \link[bnclassify]{tan_hc}.}
}
\value{
An object of class \code{DE}.
}
\description{
Learn parameters of a Bayesian Network in a discriminative way
by Algorithm jSO, an improved variant of iL-SHADE. Uses a weighted
mutation strategy (current-to-pbest-w/1) with phase-based parameter
constraints that favor exploration early and exploitation late.
}
\examples{
data(car)
dpl.jso <- jso(NP = 20, G = 25, data = car, class.name = names(car)[7], verbose = 5)
print(dpl.jso)
\dontrun{plot(dpl.jso)}
}
\references{
Brest, J., Maucec, M.S., Boskovic, B. (2017). Single Objective
Real-Parameter Optimization: Algorithm jSO. IEEE CEC 2017.
}
