% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keepSum}
\alias{keepSum}
\title{Repair to keep the sum of row vectors equal to 1
over a mutant vector}
\usage{
keepSum(x, s)
}
\arguments{
\item{x}{is the mutant vector}

\item{s}{is the netwotk structure given by \code{strRep}}
}
\description{
Repair to keep the sum of row vectors equal to 1
over a mutant vector
}
\keyword{internal}
