% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keepSumFast}
\alias{keepSumFast}
\title{Fast version of keepSum using precomputed group indices}
\usage{
keepSumFast(x, group_indices)
}
\arguments{
\item{x}{the mutant vector}

\item{group_indices}{list of precomputed index vectors for each group}
}
\description{
Fast version of keepSum using precomputed group indices
}
\keyword{internal}
