% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lev2}
\alias{lev2}
\title{Repair to keep the sum of row vectors equal to 1
In a two entry way table}
\usage{
lev2(x, index)
}
\description{
Repair to keep the sum of row vectors equal to 1
In a two entry way table
}
\keyword{internal}
