% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lev3}
\alias{lev3}
\title{Repair to keep the sum of row vectors equal to 1
In a three entry way table}
\usage{
lev3(x, index)
}
\description{
Repair to keep the sum of row vectors equal to 1
In a three entry way table
}
\keyword{internal}
