% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maswchains.R
\name{maswchains}
\alias{maswchains}
\title{Discriminative parameter learning of BN by MA-SW-Chains.}
\usage{
maswchains(
  NP = 40,
  G = 100,
  data,
  class.name,
  F = 0.5,
  CR = 0.9,
  structure = c("nb", "tancl", "tan", "hc"),
  edgelist = NULL,
  ls_evals_ratio = 0.5,
  verbose = 25,
  ...
)
}
\arguments{
\item{NP}{positive integer giving the number of candidate solutions in the initial population.}

\item{G}{positive integer specifying the maximum number of generations that may be performed before the algorithm is halted.}

\item{data}{The data frame from which to learn the classifier.}

\item{class.name}{A character. Name of the class variable.}

\item{F}{A numeric. Mutation factor for DE/rand/1. Default is 0.5.}

\item{CR}{A numeric. Cross over factor. Default is 0.9.}

\item{structure}{A character. Name of the structure learning function. "tan" uses Tree Augmented Network.
"nb" uses Naive Bayes. "hc" uses Hill Climbing.}

\item{edgelist}{A matrix. An optional edge list to use a custom BN structure
that will replace de learned structure.}

\item{ls_evals_ratio}{Numeric. Fraction of NP used as total LS budget per generation. Default is 0.5.}

\item{verbose}{positive integer indicating the number of generations until the iteration progress should be printed.}

\item{...}{other structure learning options from \link[bnclassify]{tan_cl} or \link[bnclassify]{tan_hc}.}
}
\value{
An object of class \code{DE}, which is a list with the following components:
\item{Best}{A \code{bnc_bn} object with the best individual in the final population.}
\item{BestCLL}{A numeric specifying the Conditional Log-Likelihood of the best individual.}
\item{pobFinal}{A list of \code{bnc_bn} objects with the final population.}
\item{CLLPobFinal}{A numeric vector specifying the Conditional Log-Likelihood of the final population.}
\item{N.evals}{An integer giving the total number of evaluations.}
\item{convergence}{A numeric vector giving the maximum Conditional Log-Likelihood at each generation.}
\item{evaluations}{An integer vector giving the total number of evaluations at each generation.}
}
\description{
Learn parameters of a Bayesian Network in a discriminative way
by a Memetic Algorithm with Solis-Wets Chains. Each individual maintains
its own persistent Solis-Wets state. After each DE generation, the most
promising individuals receive longer local search chains.
}
\examples{
data(car)
dpl.masw <- maswchains(NP = 20, G = 25, data = car, class.name = names(car)[7],
verbose = 5)
print(dpl.masw)
\dontrun{plot(dpl.masw)}
}
\references{
Molina, D., Lozano, M., Herrera, F. (2010). MA-SW-Chains: Memetic Algorithm
based on Local Search Chains. Evolutionary Computation 18(1).
}
