% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mos.R
\name{mos}
\alias{mos}
\title{Discriminative parameter learning of BN by MOS.}
\usage{
mos(
  NP = 40,
  G = 100,
  data,
  class.name,
  structure = c("nb", "tancl", "tan", "hc"),
  pB = 0.05,
  edgelist = NULL,
  n_splits = 10,
  min_participation = 0.1,
  verbose = 25,
  ...
)
}
\arguments{
\item{NP}{positive integer giving the number of candidate solutions in the initial population.}

\item{G}{positive integer specifying the maximum number of generations that may be performed before the algorithm is halted.}

\item{data}{The data frame from which to learn the classifier.}

\item{class.name}{A character. Name of the class variable.}

\item{structure}{A character. Name of the structure learning function. "tan" uses Tree Augmented Network.
"nb" uses Naive Bayes. "hc" uses Hill Climbing.}

\item{pB}{Numeric. Percentage of best individuals for SHADE mutation strategy. Default is 0.05.}

\item{edgelist}{A matrix. An optional edge list to use a custom BN structure
that will replace de learned structure.}

\item{n_splits}{Positive integer. Number of budget partitions. Default is 10.}

\item{min_participation}{Numeric. Minimum participation per technique. Default is 0.1.}

\item{verbose}{positive integer indicating the number of generations until the iteration progress should be printed.}

\item{...}{other structure learning options from \link[bnclassify]{tan_cl} or \link[bnclassify]{tan_hc}.}
}
\value{
An object of class \code{DE}, which is a list with the following components:
\item{Best}{A \code{bnc_bn} object with the best individual in the final population.}
\item{BestCLL}{A numeric specifying the Conditional Log-Likelihood of the best individual.}
\item{pobFinal}{A list of \code{bnc_bn} objects with the final population.}
\item{CLLPobFinal}{A numeric vector specifying the Conditional Log-Likelihood of the final population.}
\item{N.evals}{An integer giving the total number of evaluations.}
\item{convergence}{A numeric vector giving the maximum Conditional Log-Likelihood at each generation.}
\item{evaluations}{An integer vector giving the total number of evaluations at each generation.}
}
\description{
Learn parameters of a Bayesian Network in a discriminative way
by Multiple Offspring Sampling. Dynamically combines SHADE (adaptive DE)
and Solis-Wets local search, adjusting participation of each technique
based on recent performance.
}
\examples{
data(car)
dpl.mos <- mos(NP = 20, G = 25, data = car, class.name = names(car)[7],
n_splits = 5, verbose = 5)
print(dpl.mos)
\dontrun{plot(dpl.mos)}
}
\references{
LaTorre, A., Muelas, S., Pena, J.M. (2013). Large Scale Global Optimization:
Experimental Results with MOS-based Hybrid Algorithms. IEEE CEC 2013.
}
