% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mtsLS1}
\alias{mtsLS1}
\title{MTS-LS1: coordinate-wise local search}
\usage{
mtsLS1(
  x_vec,
  f_x,
  SR,
  max_dims,
  budget,
  group_indices,
  Z,
  id_params,
  offsets,
  cll_data
)
}
\arguments{
\item{x_vec}{Current parameter vector}

\item{f_x}{Current fitness (CLL) value}

\item{SR}{Step sizes per dimension}

\item{max_dims}{Maximum dimensions to perturb per call}

\item{budget}{Maximum evaluations}

\item{group_indices}{Precomputed group indices}

\item{Z}{Template bnc_bn object}

\item{id_params}{Index of .params in Z}

\item{offsets}{Precomputed offsets}

\item{cll_data}{Precomputed CLL data}
}
\description{
MTS-LS1: coordinate-wise local search
}
\keyword{internal}
