% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlshadersp.R
\name{nlshadersp}
\alias{nlshadersp}
\title{Discriminative parameter learning of BN by NL-SHADE-RSP.}
\usage{
nlshadersp(
  NP = 40,
  G = 100,
  data,
  class.name,
  c = 0.1,
  structure = c("nb", "tancl", "tan", "hc"),
  pB = 0.05,
  edgelist = NULL,
  verbose = 25,
  ...
)
}
\arguments{
\item{NP}{positive integer giving the number of candidate solutions in the initial population.}

\item{G}{positive integer specifying the maximum number of generations.}

\item{data}{The data frame from which to learn the classifier.}

\item{class.name}{A character. Name of the class variable.}

\item{c}{Numeric. An adaptation parameter. Default is 0.1.}

\item{structure}{A character. Name of the structure learning function. "tan" uses Tree Augmented Network.
"nb" uses Naive Bayes. "hc" uses Hill Climbing.}

\item{pB}{Numeric. Percentage of best individuals for mutation strategy. Default is 0.05.}

\item{edgelist}{A matrix. An optional edge list to use a custom BN structure.}

\item{verbose}{positive integer indicating the number of generations until the iteration progress should be printed.}

\item{...}{other structure learning options from \link[bnclassify]{tan_cl} or \link[bnclassify]{tan_hc}.}
}
\value{
An object of class \code{DE}.
}
\description{
Learn parameters of a Bayesian Network in a discriminative way
by NL-SHADE-RSP (Non-Linear population size reduction SHADE with
Ranked-based Selective Pressure). Uses exponential population reduction
and fitness-ranked selection of donor vectors.
}
\examples{
data(car)
dpl.nlsr <- nlshadersp(NP = 20, G = 25, data = car, class.name = names(car)[7],
c = 0.1, pB = 0.05, verbose = 5)
print(dpl.nlsr)
\dontrun{plot(dpl.nlsr)}
}
\references{
Stanovov, V., Akhmedova, S., Semenkin, E. (2021). NL-SHADE-RSP
Algorithm with Adaptive Archive and Selective Pressure for CEC 2021. IEEE CEC 2021.
}
