% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PARGEN}
\alias{PARGEN}
\title{Generate a valid conditional probability table for
attribute Xa given its parents Xb and Xc
This a three entry table}
\usage{
PARGEN(x)
}
\description{
Generate a valid conditional probability table for
attribute Xa given its parents Xb and Xc
This a three entry table
}
\keyword{internal}
