% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.DE}
\alias{plot.DE}
\title{Plot main results of evolution}
\usage{
\method{plot}{DE}(x, ...)
}
\arguments{
\item{x}{A list of structures with parameters learned by \code{\link{DE}}.}

\item{\ldots}{further arguments passed to plot.}
}
\value{
Nothing. Side-effect: plots graphs.
}
\description{
Plot main results of evolution
}
