% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{population}
\alias{population}
\title{Generates a population of potential solutions}
\usage{
population(pobSize, W, X, Y, Z)
}
\arguments{
\item{pobSize}{Number of desired individuals at initial population}
}
\description{
Generates a population of potential solutions
}
\keyword{internal}
