% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{precompute_cll_data}
\alias{precompute_cll_data}
\title{Precompute data indices for fast CLL evaluation}
\usage{
precompute_cll_data(data, Z)
}
\arguments{
\item{data}{The dataset}

\item{Z}{A bnc_bn object}
}
\description{
Precompute data indices for fast CLL evaluation
}
\keyword{internal}
