% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{randC_jso}
\alias{randC_jso}
\title{Cauchy random for jSO (scale 0.1, truncated to (0, 1])}
\usage{
randC_jso(Q, mF)
}
\description{
Cauchy random for jSO (scale 0.1, truncated to (0, 1])
}
\keyword{internal}
