% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shademts.R
\name{shademts}
\alias{shademts}
\title{Discriminative parameter learning of BN by SHADE-MTS.}
\usage{
shademts(
  NP = 40,
  G = 100,
  data,
  class.name,
  c = 0.1,
  structure = c("nb", "tancl", "tan", "hc"),
  pB = 0.05,
  edgelist = NULL,
  ls_freq = 3,
  ls_max_dims = 200,
  verbose = 25,
  ...
)
}
\arguments{
\item{NP}{positive integer giving the number of candidate solutions in the initial population.}

\item{G}{positive integer specifying the maximum number of generations.}

\item{data}{The data frame from which to learn the classifier.}

\item{class.name}{A character. Name of the class variable.}

\item{c}{Numeric. An adaptation parameter. Default is 0.1.}

\item{structure}{A character. Name of the structure learning function. "tan" uses Tree Augmented Network.
"nb" uses Naive Bayes. "hc" uses Hill Climbing.}

\item{pB}{Numeric. Percentage of best individuals for mutation strategy. Default is 0.05.}

\item{edgelist}{A matrix. An optional edge list to use a custom BN structure.}

\item{ls_freq}{Positive integer. Apply MTS-LS1 every ls_freq generations. Default is 3.}

\item{ls_max_dims}{Positive integer. Maximum dimensions to perturb per LS call. Default is 200.}

\item{verbose}{positive integer indicating the number of generations until the iteration progress should be printed.}

\item{...}{other structure learning options from \link[bnclassify]{tan_cl} or \link[bnclassify]{tan_hc}.}
}
\value{
An object of class \code{DE}.
}
\description{
Learn parameters of a Bayesian Network in a discriminative way
by SHADE hybridized with MTS-LS1 (Multiple Trajectory Search Local Search 1).
SHADE handles global exploration while MTS-LS1 performs coordinate-wise
local search on the best individual, exploiting the near-separable
structure of BN parameter spaces.
}
\examples{
data(car)
dpl.smts <- shademts(NP = 20, G = 25, data = car, class.name = names(car)[7],
c = 0.1, pB = 0.05, ls_freq = 3, verbose = 5)
print(dpl.smts)
\dontrun{plot(dpl.smts)}
}
\references{
Tseng, L.Y., Chen, C. (2008). Multiple Trajectory Search for
Large Scale Global Optimization. IEEE CEC 2008.
}
