% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{solisWetsSearch}
\alias{solisWetsSearch}
\title{Solis-Wets Local Search on BN parameter vector}
\usage{
solisWetsSearch(
  x_vec,
  f_x,
  rho,
  bias,
  n_success,
  n_fail,
  budget,
  group_indices,
  Z,
  id_params,
  offsets,
  cll_data
)
}
\arguments{
\item{x_vec}{Current parameter vector}

\item{f_x}{Current fitness (CLL) value}

\item{rho}{Search radius}

\item{bias}{Directional bias vector}

\item{n_success}{Consecutive successes count}

\item{n_fail}{Consecutive failures count}

\item{budget}{Maximum evaluations}

\item{group_indices}{Precomputed group indices for keepSumFast}

\item{Z}{Template bnc_bn object}

\item{id_params}{Index of .params in Z}

\item{offsets}{Precomputed offsets for vec2netFast}

\item{cll_data}{Precomputed CLL data}
}
\description{
Solis-Wets Local Search on BN parameter vector
}
\keyword{internal}
