% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strRep}
\alias{strRep}
\title{Repair to keep the sum of row vectors equal to 1
Gives BN structure}
\usage{
strRep(x)
}
\description{
Repair to keep the sum of row vectors equal to 1
Gives BN structure
}
\keyword{internal}
