% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec2netFast}
\alias{vec2netFast}
\title{Fast version of vec2net using precomputed offsets}
\usage{
vec2netFast(vec, Z, id_params, offsets)
}
\arguments{
\item{vec}{parameter vector}

\item{Z}{template bnc_bn object}

\item{id_params}{index of .params in Z}

\item{offsets}{precomputed list with start/end indices per node}
}
\description{
Fast version of vec2net using precomputed offsets
}
\keyword{internal}
