% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0dplbnDE-doc.R
\name{DE}
\alias{DE}
\title{Discriminative parameter learning of bayesian networks by differential evolution}
\description{
A list with Bayesian Networks, structure and parameters learned in a discriminative way by
diferential evolution returned by \code{\link{lshade}}, \code{\link{jade}},
\code{\link{DEbest}}, or \code{\link{DErand}} functions.
}
\examples{
data(car)
dpl.lshade <- lshade(NP=40, G=50, data = car, class.name = names(car)[7], c = 0.1,
structure = "tan", pB=0.05, edgelist = NULL, verbose = 5)
dpl.lshade
}
