% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lshade.R, R/methods.R
\name{plot.DE}
\alias{plot.DE}
\title{Plot main results of evolution}
\usage{
\method{plot}{DE}(x, ...)

\method{plot}{DE}(x, ...)
}
\arguments{
\item{object}{A \code{\link{DE}} object.}
}
\description{
Plot main results of evolution

Plot main results of evolution
}
