% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.DE}
\alias{print.DE}
\title{Print main results of evolution}
\usage{
\method{print}{DE}(x, ...)
}
\arguments{
\item{x}{A list of structures with parameters learned by \code{\link{DE}}.}

\item{\ldots}{Not used.}
}
\value{
Nothing. Side-effect: print results.
}
\description{
Print main results of evolution
}
\keyword{internal}
